feat: add ability to use externally allocated ASYNC_JOBs
When used this avoids the need to have a per-thread job allocation pool
and instead, for these user-allocated jobs, relies on the user to manage
these. This has the advantage of not relying on the stability of thread
local data in between context switches. I.e. the ASYNC job scheduler no
longer enforces a requirement for thread affinity in this case.

Thus if the used cryptographic ENGINE implementation also doesn't rely
on any thread-bound context it's now possible to continue a job on a
different thread from the one it last PAUSED on. This allows using I/O
event loops that don't support thread affinity of scheduled event
handlers. This should also result in better CPU utilization because
waiting on a specific thread is no longer required.